test: Replace bitrotted netbox links
Unfortunately the netbox image URLs have stopped working.
Additionally complicating this is that the mentioned netbox Git
revision is not in any released version, and the netbox project
itseelf seems inactive so no release is likely to be forthcoming.

The linked netbox fork is a faithful Git workflow build of the
previously used netbox Git revision and as a release should not
expire liek the previous nightly version did.