provider: bring back userland probes
Add uprobe and uretprobe providers for userland probing.
This functionality was originally removed in the 2.0 rewrite.

Since ply now uses the specific probe commands for reading kernel
or userland memory instead of the magic all-in-one function, we need
a way to make the IR emit the right command when dealing with
userland memory. Add a uptr(x) function for this that can mark a value
as a userland pointer, causing e.g. str(uptr(arg0)) to do the right
thing and not try to read the value as kernel memory.

Additionally we now have support for reading ELF symbol tables, so
users can specify probes such as 'uprobe:/bin/bash:readline'.