built-in: promote uptr to a built-in
Since we no longer use the magic mem_read function that would
automatically try to detect user or kernel memory, it turns out
that uptr also has utility for kprobes, e.g. do_sys_open has the
filename as a user argument so it must be read with the user
memory read function.